LayoutTests/platform/mac-wk2/fast/events/wheel/wheel-events-become-non-cancelable-expected.txt

Tests that events in the gesture are non-cancelable if preventDefault was not called on the first event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstWasCancelable is true
FAIL becameNonCancelable should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
